+++ language-server/Shared.AnZwDev.ALTools/WorkspaceCommands/ConvertObjectIdsToNamesWorkspaceCommand.cs
@@ -5,7 +5,7 @@ using System.Text;
 
 namespace AnZwDev.ALTools.WorkspaceCommands
 {
-    public class ConvertObjectIdsToNamesWorkspaceCommand : SyntaxRewriterWorkspaceCommand<ConvertObjectIdsToNamesSyntaxRewriter>
+    public class ConvertObjectIdsToNamesWorkspaceCommand : SemanticModelSyntaxRewriterWorkspaceCommand<ConvertObjectIdsToNamesSyntaxRewriter>
     {
 
         public ConvertObjectIdsToNamesWorkspaceCommand(ALDevToolsServer alDevToolsServer) : base(alDevToolsServer, "convertObjectIdsToNames")